WWLib: Make sure to use 32-bit wide types in definition

+++ GeneralsMD/Code/Libraries/Source/WWVegas/WWLib/bittype.h
@@ -43,7 +43,7 @@
 
 typedef unsigned char	uint8;
 typedef unsigned short	uint16;
-typedef unsigned long	uint32;
+typedef uint32_t	      uint32;
 typedef unsigned int    uint;
 
 typedef signed char		sint8;
@@ -54,13 +54,13 @@ typedef signed int      sint;
 typedef float				float32;
 typedef double				float64;
 
-typedef unsigned long   DWORD;
+typedef uint32_t        DWORD;
 typedef unsigned short	WORD;
 typedef unsigned char   BYTE;
 typedef int             BOOL;
 typedef unsigned short	USHORT;
 typedef const char *		LPCSTR;
 typedef unsigned int    UINT;
-typedef unsigned long   ULONG;
+typedef uint32_t        ULONG;
 
 #endif //BITTYPE_H
